[AIRFLOW-3475] Move ImportError out of models.py (#4383)
